src/app/recipes/recipes.component.spec.ts
import { Content } from '@angular/compiler/src/render3/r3_ast';
import { async, ComponentFixture, TestBed } from '@angular/core/testing';
import { HttpClientTestingModule } from '@angular/common/http/testing'
import { RecipesComponent } from './recipes.component';

describe('RecipesComponent', () => {
  let component: RecipesComponent;
  let fixture: ComponentFixture<RecipesComponent>;

  beforeEach(async(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule],
      declarations: [ RecipesComponent ]
      
    })
    .compileComponents();
  }));

  beforeEach(() => {
    fixture = TestBed.createComponent(RecipesComponent);
    component = fixture.componentInstance;
    fixture.detectChanges();
  });
  
  it('should contain trivia', async(() => {
    const fixture = TestBed.createComponent(RecipesComponent);
    fixture.detectChanges();
    const compiled = fixture.debugElement.nativeElement;
    expect(compiled.querySelector('h4').textContent).toContain(String);
  }));
  it('should contain a button for each recipe type', async(() => {
    const fixture = TestBed.createComponent(RecipesComponent);
    fixture.detectChanges();
    const compiled = fixture.debugElement.nativeElement;
    expect(compiled.querySelector('button').textContent).toContain(` All Recipes`);
  }));
  it('should contain a slideshow', async(() => {
    const fixture = TestBed.createComponent(RecipesComponent);
    fixture.detectChanges();
    const compiled = fixture.debugElement.nativeElement;
    expect(compiled.querySelector('slideshow')).toContain(Content);
  }));
});
